Add github actions to build qmlls binaries
Add a github action to build a static release build of qmlls.
There should be nightly releases that are build at night on monday,
wednesday and friday (frequency to be adjusted in the future), and
"real" releases that are built when a tag is pushed.

There is also the option to trigger the workflow "by hand" to get
nightly builds.

Nightly qmlls is versioned via qtdeclarative's short SHA,
qmlls-nightly-<short sha>, and "real releases" just follow the tag name
(as long as they have 'qmlls' and not 'nightly' in their tag name).

Add some workaround for the file stripping to have debug information in
a file separate to the qmlls binary.

You can see the github action in action at
https://github.com/samishalayel/qt5/actions/workflows/build_qmlls.yaml

Task-number: QTBUG-126405
Change-Id: I5379215062d7424544663c1519c4e296baf11fba
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>