coin: Clean up documentation builders
We have two builds we need to cover:

 1. Documentation testing, aka 'DocTests' feature

    This feature builds docs for every submodule during module
    integration, using externally provisioned doc tools.

 2. Offline documentation building, aka 'Documentation' feature

    This feature builds documentation for use in e.g. Qt Creator

For both these builds we only need a single config to handle them.
The initial configs were based on Ubuntu 22.04, but with the addition
of Ubuntu 24.04, and splitting up X11 and Wayland testing, these configs
multiplied, which is not needed. The inclusion of 'documentation' in
the various test configs was also needlessly confusing when e.g.
looking at test failure results in Grafana.

The previously named 'documentation' configs for Ubuntu have been
renamed to 'developer-build', as this is the significant difference
to the non-developer-build config.

There's still two issues with these configuration:

  1. The DocTests configuration also runs the Qt auto tests.

     The DoNotRunTests and DisableTests feature flags would
     seem to be relevant for skipping the auto-test step in
     the coin test instructions, but these are hard coded in
     coin to skip the entire test workitem.

  2. The offline doc builder builds Qt and tools from scratch

     A better approach would be to depend on another build
     and just configure and build docs, but doing so requires
     more investigation.

Change-Id: I5a041dec697424b85d3b1588cd6e77a80551d2eb
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>