coin: Skip running Qt auto tests for the documentation warning build
Take two. The qtbase TestDoc instructions should not be complete.

We only need the documentation-specific instructions to build the docs,
so instruct the test instructions in qtbase to skip the auto test.

This saves time during integration, and also avoids polluting test
results in our Grafana dashboards with test failures in a
"documentation" build.

Change-Id: Ie7114816ca35a578f01deb9946a2a6bb2e577c94
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>