Revert "Add github actions to build qmlls binaries"
This reverts commit 1f426268ad973f36e1a0265ffcfc05bccb453a86.

Reason for revert: After all, it seems to be cleaner to have a separate
repository that only contains the workflow. Also, its confusing for a
user to see that the qt5 repos only has builds of one tool instead
of the entire Qt library, and if we do add releases on github then it
would get really confusing to mix qmlls and Qt library releases together
on the same repo.

Change-Id: I17ccb1036f17fa80efab3509b8f9d2a245755958
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>