coin: Build MSVC as standalone examples in a separate build directory
The MSVC platform was not there when the rest of the platforms where
adjusted to build examples as standalone. Now that it's here, add it.

This is a pre-requisite for building examples as external projects
with MSVC.

Augments d278e682c622cf5f355edc63a503af97f079ba79

Pick-to: 6.8
Task-number: QTBUG-90820
Task-number: QTBUG-96232
Change-Id: I4a77d16e0f683f5984beeedb6e6d32707847383b
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>