Build and package Qt Gradle Plugin in CI
We want to add automation to the Qt Gradle Plugin build / packaging /
release process. The first step is to build the JARs in CI and upload
those to internal servers. To enable this we'll change the Linux
configuration to also build the binaries instead of only testing.

The "Packaging" feature is added to enable fetching the binaries after
successful integrations.

Task-number: QTTA-238
Task-number: QTBUG-142800
Change-Id: I004f1171d7c28c59fad057306c6e1c89007cfde0
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Rami Potinkara <rami.potinkara@qt.io>