Set ZLIB_ROOT on Windows with MSVC
That is required for qtopenapi builds that would be enabled in the
follow-up patch.

The zlib library is already built after commit
97e80c153803d5c466215b811b29f57fd088a308. That commit also sets the
ZLIB_PATH_{arch} environment variables. So, we only need to set
ZLIB_ROOT to point to the proper zlib installation.

This patch does not cover llvm and mingw configurations, because we
need to builds zlib using these compilers first.

Task-number: QTBUG-138968
Change-Id: I5d5ad754d13bef9fc6184b41823a42d98863991a
Reviewed-by: Toni Saario <toni.saario@qt.io>